requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asyncssh==2.21.1
atpublic==7.0.0
attrs==25.4.0
billiard==4.2.4
blinker==1.9.0
cachetools==6.2.2
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.24.10
dvc==3.64.1
dvc-data==3.16.12
dvc-http==2.32.0
dvc-objects==5.1.2
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
exceptiongroup==1.3.1
filelock==3.20.0
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.10.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.45
grandalf==0.8
gto==1.9.0
hydra-core==1.3.2
idna==3.11
iterative-telemetry==0.0.10
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kombu==5.6.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
multidict==6.7.0
narwhals==2.13.0
networkx==3.6
numpy==2.3.5
omegaconf==2.3.0
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pydot==4.0.1
pygit2==1.19.0
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
scikit-learn==1.7.2
scipy==1.16.3
scmrepo==3.5.9
seaborn==0.13.2
semver==3.0.4
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.17.0
smmap==5.0.2
sqltrie==0.11.2
streamlit==1.51.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vine==5.1.0
voluptuous==0.15.2
watchdog==6.0.0
wcwidth==0.2.14
yarl==1.22.0
zc.lockfile==4.0
